feat(#98): Mudando o service para o componente correto
